Default objectId, README fix, better JSON
Version changed to 0.0.2

If you leave the objectId blank when creating an object, the id is
inferred from the path attribute of the file being uploaded.

The readme was updated to reflect this, and to fix a typo

The JSON that is returned now returns arrays consistently where the xml
parsing wasn't.